Move InGameLobbyWatcherWrapper.java and BasicGameMenuBar.java to different packages, looking at usage, move them to packages next to classes that use them.
