+++ src/games/strategy/engine/framework/GameRunner2.java
@@ -27,7 +27,7 @@ import games.strategy.debug.ErrorConsole;
 import org.apache.commons.httpclient.HostConfiguration;
 
 import games.strategy.swing.SwingAction;
-import games.strategy.common.ui.BasicGameMenuBar;
+import games.strategy.triplea.ui.BasicGameMenuBar;
 import games.strategy.debug.ClientLogger;
 import games.strategy.engine.ClientContext;
 import games.strategy.engine.ClientFileSystemHelper;
