+++ src/games/strategy/engine/framework/headlessGameServer/HeadlessServerSetup.java
@@ -7,7 +7,7 @@ import java.util.concurrent.CopyOnWriteArrayList;
 
 import javax.swing.Action;
 
-import games.strategy.common.ui.InGameLobbyWatcherWrapper;
+import games.strategy.engine.framework.startup.ui.InGameLobbyWatcherWrapper;
 import games.strategy.engine.chat.IChatPanel;
 import games.strategy.engine.data.GameData;
 import games.strategy.engine.framework.IGame;
