+++ src/games/strategy/engine/framework/networkMaintenance/SetPasswordAction.java
@@ -10,7 +10,7 @@ import javax.swing.JOptionPane;
 import javax.swing.JPanel;
 import javax.swing.JPasswordField;
 
-import games.strategy.common.ui.InGameLobbyWatcherWrapper;
+import games.strategy.engine.framework.startup.ui.InGameLobbyWatcherWrapper;
 import games.strategy.engine.framework.startup.login.ClientLoginValidator;
 
 public class SetPasswordAction extends AbstractAction {
