+++ src/games/strategy/engine/framework/startup/ui/ServerSetupPanel.java
@@ -25,7 +25,6 @@ import javax.swing.JTextField;
 import javax.swing.ScrollPaneConstants;
 import javax.swing.SwingUtilities;
 
-import games.strategy.common.ui.InGameLobbyWatcherWrapper;
 import games.strategy.engine.chat.IChatPanel;
 import games.strategy.engine.data.GameData;
 import games.strategy.engine.data.PlayerID;
