+++ src/games/strategy/triplea/ui/TripleAFrame.java
@@ -86,7 +86,6 @@ import com.google.common.collect.ImmutableList;
 
 import games.strategy.common.delegate.BaseEditDelegate;
 import games.strategy.swing.SwingAction;
-import games.strategy.common.ui.BasicGameMenuBar;
 import games.strategy.common.ui.MacWrapper;
 import games.strategy.common.ui.MainGameFrame;
 import games.strategy.debug.ClientLogger;
